+++ shared/src/task_types.rs
@@ -5,10 +5,8 @@ use serde::{Deserialize, Serialize};
 pub type FileCopyInner = (String, String);
 
 /// Represents inner data for the [`AdminCommand::BuildAllBins`], as a tuple for:
-/// (`profile_disk_name`, `save path`, `listener_profile`, `implant_profile`).
-///
-/// For `listener_profile` & `implant_profile`, a value of `None` will resolve to matching on `default`.
-pub type BuildAllBins = (String, String, Option<String>, Option<String>);
+/// (`profile_disk_name`, `save path`, `implant_profile`).
+pub type BuildAllBins = (String, String, String);
 
 pub type RegQueryInner = (String, Option<String>);